+++ src/main/java/net/brdle/collectorsreap/common/effect/VolatilityEffect.java
@@ -4,7 +4,6 @@ import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
 
 public class VolatilityEffect extends MobEffect {
-
 	/**
 	 * The player will emit acid particles that destroy incoming projectiles
 	 * Used weapon will take extra durability damage
@@ -17,7 +16,7 @@ public class VolatilityEffect extends MobEffect {
 	 * Checks whether the effect is ready to be applied this tick.
 	 */
 	@Override
-	public boolean isDurationEffectTick(int pDuration, int pAmplifier) {
+	public boolean isDurationEffectTick(int duration, int amplifier) {
 		return false;
 	}
 }