+++ src/main/java/net/brdle/collectorsreap/data/CRItemTags.java
@@ -23,8 +23,8 @@ public class CRItemTags {
 	public static final TagKey<Item> TEA_LEAVES_YELLOW = forge("tea_leaves/yellow");
 	public static final TagKey<Item> TEA_LEAVES_BLACK = forge("tea_leaves/black");
 	public static final TagKey<Item> COFFEE_BEANS = forge("coffee_beans");
-	public static final TagKey<Item> CRAB_MEAT = forge("crab_meat");
-	public static final TagKey<Item> CRAB_LEG = forge("crab_leg");
+	public static final TagKey<Item> RAW_CRAB_MEAT = forge("raw_crab_meat");
+	public static final TagKey<Item> COOKED_CRAB_MEAT = forge("cooked_crab_meat");
 	public static final TagKey<Item> RAW_CLAM = forge("raw_clam");
 	public static final TagKey<Item> RAW_PRAWN = forge("raw_prawn");
 	public static final TagKey<Item> COOKED_PRAWN = forge("cooked_prawn");